chat_sentiment_analysis/requirements.txt
requests
pandas
emoji
leia-br
pymongo
vaderSentiment
aws-lambda-powertools[tracer]
